reorganize some files in conf.d and plugins
vivid is back as an env, and is in .zshenv now.
The gain as a plugin was meaningless, and LS_COLORS may be used during the initialisation.

conf.d/autosuggestions.zsh was put back in conf.d/__init__.zsh

conf.d/zoxide.zsh was moved as a plugin, I wanted to test something being loaded after it.
And this is also to avoid loading it twice (I was loading the official plugin)

fix fzf-rg not using GREP_EXCLUDE, and only show dirs during cd completion.

new function, ssh-cp-id: copy your id_rsa.pub to a remote server